Implemented user input testing
